refactor: 엔티티 사용되지 않는 로직 제거 및 패키지 구조 용도에 맞게 변경
