+++ commerce/src/main/java/store/_0982/commerce/application/order/OrderService.java
@@ -1,59 +1,42 @@
 package store._0982.commerce.application.order;
 
-import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import store._0982.commerce.application.cart.CartService;
 import store._0982.commerce.application.grouppurchase.GroupPurchaseService;
-import store._0982.commerce.application.grouppurchase.ParticipateService;
 import store._0982.commerce.application.order.dto.*;
 import store._0982.commerce.application.order.event.OrderCancelProcessedEvent;
-import store._0982.commerce.application.order.event.OrderCartCompletedEvent;
 import store._0982.commerce.application.sellerbalance.SellerBalanceService;
-import store._0982.commerce.domain.cart.Cart;
 import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.order.Order;
 import store._0982.commerce.domain.order.OrderRepository;
 import store._0982.commerce.domain.order.OrderStatus;
 import store._0982.commerce.exception.CustomErrorCode;
-import store._0982.commerce.infrastructure.client.member.MemberClient;
-import store._0982.commerce.infrastructure.client.member.dto.ProfileInfo;
-import store._0982.commerce.infrastructure.client.payment.PaymentClient;
-import store._0982.commerce.infrastructure.client.payment.dto.PointReturnRequest;
 import store._0982.common.dto.PageResponse;
-import store._0982.common.dto.ResponseDto;
 import store._0982.common.exception.CustomException;
-import store._0982.common.log.ServiceLog;
 
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.List;
+import java.util.UUID;
 
 import static store._0982.commerce.domain.order.OrderCancellationPolicy.*;
 
 @Slf4j
 @Service
-@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class OrderService {
 
+    private final OrderCommandService orderCommandService;
+    private final OrderQueryService orderQueryService;
     private final OrderRepository orderRepository;
-    private final CartService cartService;
 
     private final GroupPurchaseService groupPurchaseService;
     private final SellerBalanceService sellerBalanceService;
-    private final ParticipateService participateService;
-
-    private final MemberClient memberClient;
-    private final PaymentClient paymentClient;
 
     private final ApplicationEventPublisher eventPublisher;
+
 
     /**
      * 주문 생성
@@ -62,42 +45,8 @@ public class OrderService {
      * @param command  주문 command
      * @return OrderRegisterInfo
      */
-    @ServiceLog
-    @Transactional
     public OrderRegisterInfo createOrder(UUID memberId, OrderRegisterCommand command) {
-
-        // 이미 처리 요청된 주문인지 확인
-        if(orderRepository.existsByIdempotencyKey(command.requestId())){
-            throw new CustomException(CustomErrorCode.DUPLICATE_ORDER);
-        }
-
-        // 주문자 존재 여부
-        String memberName = validateMember(memberId);
-
-        // 공동 구매 존재 여부
-        GroupPurchase groupPurchase = validateGroupPurchase(command.groupPurchaseId());
-
-        // 참여
-        participateService.participate(groupPurchase.getGroupPurchaseId(), command.quantity(), memberName, command.requestId());
-
-        // order 생성
-        Order order = Order.create(
-                command.quantity(),
-                groupPurchase.getDiscountedPrice(),
-                memberId,
-                command.address(),
-                command.addressDetail(),
-                command.postalCode(),
-                command.receiverName(),
-                command.sellerId(),
-                groupPurchase.getGroupPurchaseId(),
-                command.requestId()
-        );
-
-        Order savedOrder = orderRepository.save(order);
-
-        return OrderRegisterInfo.from(savedOrder);
-
+        return orderCommandService.createOrder(memberId, command);
     }
 
     /**
@@ -107,85 +56,8 @@ public class OrderService {
      * @param command  주문 command
      * @return OrderRegisterInfo List
      */
-    @Transactional
-    @ServiceLog
     public List<OrderRegisterInfo> createOrderCart(UUID memberId, OrderCartRegisterCommand command) {
-        // 주문자 존재 여부
-        String memberName = validateMember(memberId);
-
-        // cartId 리스트로 장바구니 아이템들 조회
-        List<Cart> carts = cartService.validateAndGetCartForOrder(memberId,command.cartIds());
-
-        // 공동 구매 유효한지 확인
-        Set<UUID> groupPurchaseIds = carts.stream()
-                .map(Cart::getGroupPurchaseId)
-                .collect(Collectors.toSet());
-
-        // 공동 구매 리스트 조회
-        Map<UUID, GroupPurchase> purchasesMap = groupPurchaseService.getAvailableGroupPurchasesOrder(groupPurchaseIds);
-
-        // 주문 생성
-        List<OrderRegisterInfo> orders = createOrderFromCart(memberId, carts, purchasesMap, command, memberName);
-
-        // 장바구니 비우기
-        eventPublisher.publishEvent(new OrderCartCompletedEvent(carts));
-
-        return orders;
-    }
-
-
-    private String validateMember(UUID memberId) {
-        try {
-            ResponseDto<ProfileInfo> member = memberClient.getMember(memberId);
-            return member.data().name();
-        } catch (FeignException.NotFound e) {
-            throw new CustomException(CustomErrorCode.MEMBER_NOT_FOUND);
-        } catch (FeignException e) {
-            log.error("회원 조회 실패 : memberId={}, status={}", memberId, e.status());
-            throw new RuntimeException("회원 정보 조회 오류");
-        }
-    }
-
-    private GroupPurchase validateGroupPurchase(UUID groupPurchaseId) {
-        return groupPurchaseService.getAvailableForOrder(groupPurchaseId);
-    }
-
-    private List<OrderRegisterInfo> createOrderFromCart(UUID memberId, List<Cart> carts, Map<UUID, GroupPurchase> purchaseMap, OrderCartRegisterCommand command, String memberName){
-        List<Order> orderToSave = new ArrayList<>();
-
-        for(Cart cart: carts){
-            GroupPurchase groupPurchase = purchaseMap.get(cart.getGroupPurchaseId());
-
-            if(groupPurchase == null){
-                throw new CustomException(CustomErrorCode.GROUPPURCHASE_NOT_FOUND);
-            }
-
-            String orderRequestId = command.requestId() + "-" + cart.getCartId();
-
-
-            participateService.participate(cart.getGroupPurchaseId(), cart.getQuantity(), memberName, orderRequestId);
-
-            Order order = Order.create(
-                    cart.getQuantity(),
-                    groupPurchase.getDiscountedPrice(),
-                    memberId,
-                    command.address(),
-                    command.addressDetail(),
-                    command.postalCode(),
-                    command.receiverName(),
-                    groupPurchase.getSellerId(),
-                    groupPurchase.getGroupPurchaseId(),
-                    orderRequestId
-            );
-
-            orderToSave.add(order);
-        }
-
-        List<Order> savedOrders = orderRepository.saveAll(orderToSave);
-
-        return savedOrders.stream()
-                .map(OrderRegisterInfo::from)
-                .collect(Collectors.toList());
+        return orderCommandService.createOrderCart(memberId, command);
     }
 
     /**
@@ -196,15 +68,7 @@ public class OrderService {
      * @return OrderDetailInfo
      */
     public OrderDetailInfo getOrderById(UUID requesterID, UUID orderId) {
-        Order order = orderRepository.findByOrderIdAndDeletedAtIsNull(orderId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.ORDER_NOT_FOUND));
-        log.info("getOrderById - orderId: {}", orderId);
-        if (!order.getMemberId().equals(requesterID)
-                && !order.getSellerId().equals(requesterID)) {
-            throw new CustomException(CustomErrorCode.ORDER_ACCESS_DENIED);
-        }
-
-        return OrderDetailInfo.from(order);
+        return orderQueryService.getOrderById(requesterID, orderId);
     }
 
     /**
@@ -215,16 +79,7 @@ public class OrderService {
      * @return OrderInfo
      */
     public PageResponse<OrderInfo> getOrdersBySeller(UUID sellerId, Pageable pageable) {
-
-        Pageable sortPageable = PageRequest.of(
-                pageable.getPageNumber(),
-                pageable.getPageSize(),
-                Sort.by(Sort.Direction.ASC, "createdAt"));
-
-        Page<Order> orders = orderRepository.findBySellerIdAndDeletedIsNull(sellerId, sortPageable);
-
-        Page<OrderInfo> orderInfos = orders.map(OrderInfo::from);
-        return PageResponse.from(orderInfos);
+        return orderQueryService.getOrdersBySeller(sellerId, pageable);
     }
 
     /**
@@ -235,82 +90,7 @@ public class OrderService {
      * @return OrderInfo
      */
     public PageResponse<OrderInfo> getOrdersByConsumer(UUID memberId, Pageable pageable) {
-        Page<Order> orders = orderRepository.findByMemberIdAndDeletedIsNull(memberId, pageable);
-
-        Page<OrderInfo> orderInfos = orders.map(OrderInfo::from);
-        return PageResponse.from(orderInfos);
-    }
-
-    /**
-     * 주문 상태 변경
-     *
-     * @param groupPurchaseId 공동구매 id
-     * @param orderStatus     주문 상태
-     */
-    @Transactional
-    public void updateOrderStatus(UUID groupPurchaseId, OrderStatus orderStatus) {
-        List<Order> orders = orderRepository.findByGroupPurchaseIdAndDeletedAtIsNull(groupPurchaseId);
-
-        log.info("orderStatus : {}", orderStatus);
-
-        for (Order order : orders) {
-
-            log.info("orderId : {}", order.getOrderId());
-            order.updateStatus(orderStatus);
-        }
-
-        orderRepository.saveAll(orders);
-    }
-
-    /**
-     * 주문 환불
-     *
-     * @param groupPurchaseId
-     */
-    @ServiceLog
-    @Transactional
-    public void returnOrder(UUID groupPurchaseId) {
-        List<Order> orders = orderRepository.findByGroupPurchaseIdAndStatusAndDeletedAtIsNull(groupPurchaseId, OrderStatus.GROUP_PURCHASE_FAIL);
-
-        log.info("환불 대상 주문 조회 완료 : groupPurchaseId = {}, count = {}", groupPurchaseId, orders.size());
-
-        List<Order> toUpdate = new ArrayList<>();
-
-        int success = 0;
-        int fail = 0;
-
-        for (Order order : orders) {
-
-            if (order.isReturned()) {
-                log.info("이미 환불된 주문 : {}", order.getOrderId());
-                continue;
-            }
-            Long amount = order.getPrice() * order.getQuantity();
-
-            try {
-                paymentClient.returnPointsInternal(
-                        order.getMemberId(),
-                        new PointReturnRequest(
-                                UUID.randomUUID(),
-                                order.getOrderId(),
-                                amount
-                        )
-                );
-                order.markReturned();
-                toUpdate.add(order);
-                success++;
-                log.info("환불 완료 : orderId = {}, memberId = {}, amount = {}", order.getOrderId(), order.getMemberId(), amount);
-            } catch (Exception e) {
-                fail++;
-                log.error("환불 실패 : orderId = {}, memberId = {}, amount = {}", order.getOrderId(), order.getMemberId(), amount);
-                // TODO : 환불 실패 처리 필요(재시도 큐? 관리자 알람?)
-            }
-        }
-
-        if (!toUpdate.isEmpty()) {
-            orderRepository.saveAll(toUpdate);
-            log.info("환불 상태 업데이트 완료 : 성공 = {}, 실패 = {}", success, fail);
-        }
+        return orderQueryService.getOrdersByConsumer(memberId, pageable);
     }
 
     @Transactional
@@ -380,6 +160,6 @@ public class OrderService {
     }
 
     public List<Order> getAllOrderByMemberId(UUID memberId) {
-        return orderRepository.findAllByMemberId(memberId);
+        return orderQueryService.getAllOrderByMemberId(memberId);
     }
 }
