+++ commerce/src/main/java/store/_0982/commerce/domain/order/Order.java
@@ -60,12 +60,6 @@ public class Order {
     @Column(name = "payment_method")
     private PaymentMethod paymentMethod;
 
-    @Column(name = "expired_at")
-    private OffsetDateTime expiredAt;
-
-    @Column(name = "paid_at")
-    private OffsetDateTime paidAt;
-
     @Column(name = "created_at", nullable = false)
     @CreationTimestamp
     private OffsetDateTime createdAt;
@@ -73,6 +67,12 @@ public class Order {
     @Column(name = "updated_at")
     @UpdateTimestamp
     private OffsetDateTime updatedAt;
+
+    @Column(name = "expired_at")
+    private OffsetDateTime expiredAt;
+
+    @Column(name = "paid_at")
+    private OffsetDateTime paidAt;
 
     @Column(name = "deleted_at")
     private OffsetDateTime deletedAt;
@@ -130,29 +130,33 @@ public class Order {
         );
     }
 
-    // 상태 변경
-    public void updateStatus(OrderStatus newStatus) {
-        this.status = newStatus;
-    }
 
     // 결제 완료
     public void completePayment(PaymentMethod paymentMethod) {
-        validateStatus(OrderStatus.PAYMENT_COMPLETED);
+        // 이미 결제 완료된 건인지 확인
+        if(this.status == OrderStatus.PAYMENT_COMPLETED){
+            return;
+        }
+
+        if(this.status != OrderStatus.PENDING){
+            throw new CustomException(CustomErrorCode.CANNOT_PAYMENT_COMPLETED_ORDER_INVALID_STATUS);
+        }
         this.status = OrderStatus.PAYMENT_COMPLETED;
         this.paymentMethod = paymentMethod;
         this.paidAt = OffsetDateTime.now();
     }
 
-    // 주문 실패 처리
-    public void markFailed() {
-        validateStatus(OrderStatus.ORDER_FAILED);
-        this.status = OrderStatus.ORDER_FAILED;
-    }
-
     // 주문 취소 처리
-    public void cancel() {
-        validateStatus(OrderStatus.CANCELLED);
-        this.status = OrderStatus.CANCELLED;
+    public void markFailed() {
+        // 이미 결제 실패된 건인지 확인
+        if(this.status == OrderStatus.PAYMENT_FAILED){
+            return;
+        }
+
+        if(this.status != OrderStatus.PENDING){
+            throw new CustomException(CustomErrorCode.CANNOT_PAYMENT_FAILED_ORDER_INVALID_STATUS);
+        }
+        this.status = OrderStatus.ORDER_FAILED;
     }
 
     public void requestCancel() {
@@ -164,39 +168,13 @@ public class Order {
     public void requestReversed() {
         if (this.status != OrderStatus.GROUP_PURCHASE_SUCCESS)
             throw new CustomException(CustomErrorCode.CANNOT_REVERSE_ORDER_INVALID_STATUS);
-        this.status = OrderStatus.REVERSED_REQUESTED;
+        this.status = OrderStatus.REVERSE_REQUESTED;
     }
 
     public void requestReturned() {
         if (this.status != OrderStatus.GROUP_PURCHASE_SUCCESS)
             throw new CustomException(CustomErrorCode.CANNOT_RETURN_ORDER_INVALID_STATUS);
-        this.status = OrderStatus.REVERSED_REQUESTED;
-    }
-
-    private void validateStatus(OrderStatus newStatus) {
-        switch (this.status) {
-            case PENDING:
-                if (newStatus != OrderStatus.PAYMENT_COMPLETED
-                        && newStatus != OrderStatus.CANCELLED
-                        && newStatus != OrderStatus.ORDER_FAILED) {
-                    throw new IllegalStateException("PENDING으로 변경 불가능");
-                }
-                break;
-            case PAYMENT_COMPLETED:
-                if (newStatus != OrderStatus.CANCELLED
-                        && newStatus != OrderStatus.REFUNDED
-                        && newStatus != OrderStatus.REVERSED) {
-                    throw new IllegalStateException("PAYMENT_COMPLETED로 변경 불가능");
-                }
-                break;
-            case ORDER_FAILED:
-            case CANCELLED:
-            case REFUNDED:
-                throw new IllegalStateException("상태 변경 불가능");
-
-            default:
-                break;
-        }
+        this.status = OrderStatus.REVERSE_REQUESTED;
     }
 
     // 환불 완료
