+++ commerce/src/main/java/store/_0982/commerce/infrastructure/kafka/publisher/OrderKafkaEventPublisher.java
@@ -1,0 +1,4 @@
+package store._0982.commerce.infrastructure.kafka.publisher;
+
+public class OrderKafkaEventPublisher {
+}
