+++ commerce/src/main/resources/db/migration/V4_1__drop_settled_at.sql
@@ -1,0 +1,3 @@
+ALTER TABLE product_schema.group_purchase
+DROP
+COLUMN settled_at;
